// src/types/Courses.d.ts
export interface Course {
    courseTitleLong: string;
    courseDescription: string;
    classSection: string;
    enrollmentStatus: string;
    minimumUnits: number;
    instructorFullName: string;
    instructorEmail: string;
    enrollmentTotal: number;
    enrollmentCap: number;
    waitTotal: number;
    waitCap: number;
}